search: code monitoring, move queries to store
This is a pure refactor. This PR moves all db-interactions with
cm_queries from resolver.go to to the store (../queries.go).
